+++ backend/Models/Coin.cs
@@ -3,14 +3,14 @@ public record Coin
 {
     public string Name { get; init; }
     public string TagName { get; init; }
-    public double Value { get; init; }
-    public double Weight { get; init; }  
+    public decimal Value { get; init; }
+    public decimal Weight { get; init; }
 
-    public Coin(string name, string tagName,  double value, double weight)
+    public Coin(string name, string tagName, decimal value, decimal weight)
     {
         Name = name ?? throw new ArgumentNullException(nameof(name));
         TagName = tagName ?? throw new ArgumentNullException(nameof(tagName));
         Value = value;
         Weight = weight;
-    }    
+    }
 }
